create a single table
