+++ conf/modules.config
@@ -53,15 +53,7 @@ process {
         ext.args = "--report sequence"
     }
 
-    withName: SUMMARYTABLE {
-        publishDir = [
-            path: { "${meta.outdir}/genome_statistics" },
-            mode: params.publish_dir_mode,
-            saveAs: { filename -> filename.equals('versions.yml') ? null : filename }
-        ]
-    }
-
-    withName: ADDMERQURY {
+    withName: CREATETABLE {
         publishDir = [
             path: { "${meta.outdir}/genome_statistics" },
             mode: params.publish_dir_mode,
