+++ modules/local/createtable.nf
@@ -1,4 +1,4 @@
-process SUMMARYTABLE {
+process CREATETABLE {
     tag "$meta.id"
     label 'process_single'
 
@@ -8,7 +8,9 @@ process SUMMARYTABLE {
         'quay.io/biocontainers/python:3.9--1' }"
 
     input:
-    tuple val(meta), path(genome_summary), path(sequence_summary), path(busco)
+    tuple val(meta), path(genome_summary), path(sequence_summary)
+    path(busco)
+    tuple val(meta2), path(qv), path(completeness)
 
     output:
     tuple val(meta), path("*.csv"), emit: csv
@@ -19,16 +21,25 @@ process SUMMARYTABLE {
 
     script: // This script is bundled with the pipeline, in sanger-tol/genomenote/bin/
     def prefix = task.ext.prefix ?: "${meta.id}"
+    def gen = genome_summary ? "--genome ${genome_summary}" : ""
+    def seq = sequence_summary ? "--sequence ${sequence_summary}" : ""
+    def bus = busco ? "--busco ${busco}" : ""
+    def pac = qv ? "--pacbio ${meta2.id}" : ""
+    def mqv = qv ? "--qv ${qv}" : ""
+    def mco = completeness ? "--completeness ${completeness}" : ""
     """
-    summary_table.py \\
-        $genome_summary \\
-        $sequence_summary \\
-        $busco \\
-        ${prefix}.csv
+    create_table.py \\
+        $gen \\
+        $seq \\
+        $bus \\
+        $pac \\
+        $mqv \\
+        $mco \\
+        -o ${prefix}.csv
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
-        summary_table.py: \$(summary_table.py --version | cut -d' ' -f2)
+        create_table.py: \$(summary_table.py --version | cut -d' ' -f2)
     END_VERSIONS
     """
 }
